refactor: move screenshot file validation logic to utils and clean up endpoints
